pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "AUR"
version = "2021.11.20"
description = "AUR-related modules and helper utilities (aurploader, aurquery, aurtomatic)."
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Xyne", email = "xyne@archlinux.org"},
]
dependencies = [
    "pyalpm",
    "pyxdg",
    # https://aur.archlinux.org/packages/python3-xcgf
    "xcgf",
    # https://aur.archlinux.org/packages/python3-xcpf
    "xcpf"
]

[project.urls]
Home = "https://xyne.dev/projects/python3-aur"

[project.scripts]
aurpkglist = "aurpkglist"
aurploder = "aurploder"
aurquery = "aurquery"
aurtomatic = "aurtomatic"
aurtus = "aurtus"
